requirements.txt
line-bot-sdk
flask
requests==1.1.0
python-firebase